+++ src/components/Dashboards/Dashboard.module.css
@@ -1,4 +1,4 @@
-.admin {
+.dashboard {
     display: flex;
     flex-direction: column;
     background: rgb(236, 238, 240);
@@ -94,4 +94,12 @@
     margin-right: 50px;
 }
 
+.top{
+    display: inline-flex !important;
+}
+
+.pagination{
+    float:right;
+    margin-bottom: 50px;
+}
